* networking actions are now associated with ports rather than nics.
Without this patch we cannot perform networking actions on a port that's not
connected to a nic. This prevents us from calling port revert as a part of port
initialization. With this patch, we can now add support for setting jumbo
frames, or changing STP settings as part of port initialization.